skip only if correct verison

+++ package/opt/hassbian/suites/python.sh
@@ -26,22 +26,25 @@ fi
 
 isinstalled=$(which python${PYTHONVERSION:0:3})
 if [ ! -z "${isinstalled}" ]; then
-  echo "Upgrading to Python $PYTHONVERSION"
-  apt-get -y update
-  apt-get install -y build-essential tk-dev libncurses5-dev libncursesw5-dev libreadline6-dev libdb5.3-dev libgdbm-dev libsqlite3-dev libssl-dev libbz2-dev libexpat1-dev liblzma-dev zlib1g-dev
-  apt-get install libtcmalloc-minimal4
-  export LD_PRELOAD="/usr/lib/libtcmalloc_minimal.so.4"
-  cd /tmp || return 1
-  wget https://www.python.org/ftp/python/"$PYTHONVERSION"/Python-"$PYTHONVERSION".tar.xz
-  tar xf Python-"$PYTHONVERSION".tar.xz
-  cd Python-"$PYTHONVERSION" || return 1
-  ./configure
-  make altinstall
-  apt -y autoremove
-  cd  || return 1
-  rm -r /tmp/Python-"$PYTHONVERSION"
-  rm /tmp/Python-"$PYTHONVERSION".tar.xz
-  echo "Done"
+  currentversion$(python${PYTHONVERSION:0:3} -V)
+  if [[ "$currentversion" != "$PYTHONVERSION" ]]; then
+    echo "Upgrading to Python $PYTHONVERSION"
+    apt-get -y update
+    apt-get install -y build-essential tk-dev libncurses5-dev libncursesw5-dev libreadline6-dev libdb5.3-dev libgdbm-dev libsqlite3-dev libssl-dev libbz2-dev libexpat1-dev liblzma-dev zlib1g-dev
+    apt-get install libtcmalloc-minimal4
+    export LD_PRELOAD="/usr/lib/libtcmalloc_minimal.so.4"
+    cd /tmp || return 1
+    wget https://www.python.org/ftp/python/"$PYTHONVERSION"/Python-"$PYTHONVERSION".tar.xz
+    tar xf Python-"$PYTHONVERSION".tar.xz
+    cd Python-"$PYTHONVERSION" || return 1
+    ./configure
+    make altinstall
+    apt -y autoremove
+    cd  || return 1
+    rm -r /tmp/Python-"$PYTHONVERSION"
+    rm /tmp/Python-"$PYTHONVERSION".tar.xz
+    echo "Done"
+  fi
 fi
 echo "Stopping Home Assistant."
 systemctl stop home-assistant@homeassistant.service
